Novo package de testes criado
