+++ src/gerenciamento/hotel/GerenciamentoHotel.java
@@ -5,10 +5,7 @@
  */
 package gerenciamento.hotel;
 
-import estacionamento.*;
-import veiculos.*;
-import java.util.ArrayList;
-import java.util.Arrays;
+import Testes.Estacionamento;
 
 /**
  *
@@ -20,30 +17,8 @@ public class GerenciamentoHotel {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        Bicicleta bike =  new Bicicleta("preto");
-        Motocicleta moto = new Motocicleta("Verde", 2);
-        moto.setCilindrada(200);
-        Carro carro = new Carro("Cinza", 5);
-        carro.setAno(2012);
-        carro.setModelo("Uno Fiat");
-        
-        Vaga v1 = new Vaga(1);
-        v1.addBicicleta(bike);
-        v1.addMotocicleta(moto);
-        
-        Vaga v2 = new Vaga(2);
-        v2.addCarro(carro);
-        
-        ArrayList<Vaga> vagas = new ArrayList(Arrays.asList(v1,v2));
-        
-        Estacionamento estacionamento =  new Estacionamento(vagas);
-        
-        estacionamento.addVaga(v1);
-        estacionamento.addVaga(v2);
-        
-        estacionamento.listVagas();
-        
-        
+        // Roda seeds do estacionamento (Veiculos, Vaga)
+        new Estacionamento();
     }
     
 }
