--- test/Estacionamento.java
@@ -1,13 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-/**
- *
- * @author matheusfaggi
- */
-public class Estacionamento {
-    
-}
